+++ CHANGELOG.rst
@@ -11,6 +11,13 @@ Changelog
 * Parsing SSH certificates no longer permits malformed critical options with
   values, as documented in the 41.0.2 release notes.
 * Updated the minimum supported Rust version (MSRV) to 1.63.0, from 1.56.0.
+* Support :class:`~cryptography.hazmat.primitives.asymmetric.padding.PSS` for
+  X.509 certificate signing requests with the keyword-only argument
+  ``rsa_padding`` on
+  :meth:`~cryptography.x509.CertificateSigningRequestBuilder.sign`.
+* Added support for obtaining X.509 certificate signing request signature
+  algorithm parameters (including PSS) via
+  :meth:`~cryptography.x509.CertificateSigningRequest.signature_algorithm_parameters`.
 
 .. _v41-0-3: